==> 02_Exploratory_Data_Analysis.Rmd <==
---
title: "02_Exploratory_Data_Analysis"
date: '`r Sys.Date()`'
output:
  bookdown::html_document2: default
---

```{r library, echo = FALSE, include = FALSE}

library(dplyr)
library(tidyverse)
library(ggplot2)
library(knitr)
library(kableExtra)
library(corrplot)
library(ggcorrplot) 
library(gridExtra)
library(bookdown)
library(moments)
library(here)
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
theme_set(theme_minimal())
```
::: {style="text-align: justify;"}

# Exploratory Data Analysis (EDA)

This document presents an exploratory data analysis of critical elements concentration data, which provide information about the concentration of critical mineral across different areas in Australia that has been normalised using PAAS (Post-Archean Australian Shale) standard. The primary objective of this analysis is to gain a deeper understanding of the data's structure and key characteristics. Through this, we aim to identify significant trends, correlations, and outliers that may influence the outcomes of the study.

<br />

## Data General Characteristics
```{r}
# Loading all data
alldata <- readRDS(here('results/all_data.rds'))

# Exclude observations that element value below -800.00 data
# alldata_eda <- alldata[alldata$Element_Value_ppm >= -800.00, ]
```
Our data comprises of **`r nrow(alldata)` observations** and **`r ncol(alldata)` variables**. Most of the variables are character type, except for variable `Element_Value_ppm`, `PAAS_value_ppm`, and `PAAS_normalised_value` that are numeric type.

```{r}
str(alldata)
```


The **Table \@ref(tab:dat-preview)** provide a preview of the data from the first ten observations.

```{r dat-preview, fig.cap = "First ten observations of the data"}
kable(head(alldata, n = 10),
      col.names = gsub("[_]", " ", names(alldata)),
      align = "llcccccl",
      caption = "The first ten observations of the data",
      format = "html") %>%
  kableExtra::kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"),
                            full_width = F,
                            font_size = 14,
                            fixed_thead = T)
```

<br />

## Descriptive Statistics

In this section, we will provide some generic statistics information about the data, such as min, max, mean, median, and others. The **Table \@ref(tab:desc-stats)** lay out the details about these informations for each critical element that is included in the data.
```{r desc-stats, out.width = "50%", fig.align = "center"}
# Populate descriptive stats table
desc_stats <- alldata %>%
  group_by(Element_Symbol, Element_Description) %>%
  summarise(min = min(Element_Value_ppm),
         max = max(Element_Value_ppm),
         mean = mean(Element_Value_ppm),
         median = median(Element_Value_ppm),
         range = max(Element_Value_ppm) - min(Element_Value_ppm),
         q1 = quantile(Element_Value_ppm, 0.25),
         q3 = quantile(Element_Value_ppm, 0.75),
         iqr = IQR(Element_Value_ppm),
         sd = sd(Element_Value_ppm),
         var = var(Element_Value_ppm),
         skewness = moments::skewness(Element_Value_ppm),
         kurtosis = moments::kurtosis(Element_Value_ppm))

knitr::kable(desc_stats, 
             col.names = gsub("[_]", " ", names(desc_stats)),
             digits = 3, 
             format.args = list(big.mark = ",", scientific = FALSE),
             format = "html", 
             align = "c", 
             caption = "Descriptive Statistics of Elements' Normalised Value") %>%
  kableExtra::kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"),
                            full_width = F,
                            font_size = 14,
                            fixed_thead = T) %>%
  kableExtra::column_spec(2, 
                          extra_css = "max-width: 100px; word-wrap: break-word; overflow-wrap: break-word;") %>%
  scroll_box(height = "500px",
             extra_css = "border: none; box-shadow: none; padding: 5px; ")
```

The most important statistics from this table is kurtosis. Kurtosis measures the combined weight of the tails of a distribution relative to its centre. In this way, we can use kurtosis as an indicator of the presence of outliers. A high kurtosis values is indicative of outliers. Validating the outliers will be easier with data visualisation, which will be presented in the next section.

The table below represents the descriptive statistics of elements from ME-4ACD81 test that will be used for predictive modelling.
```{r}
# Populate descriptive stats table
desc_stats_flt <- alldata %>%
  group_by(Element_Symbol, Element_Description) %>%
  summarise(min = min(Element_Value_ppm),
         max = max(Element_Value_ppm),
         mean = mean(Element_Value_ppm),
         median = median(Element_Value_ppm),
         range = max(Element_Value_ppm) - min(Element_Value_ppm),
         q1 = quantile(Element_Value_ppm, 0.25),
         q3 = quantile(Element_Value_ppm, 0.75),
         iqr = IQR(Element_Value_ppm),
         sd = sd(Element_Value_ppm),
         var = var(Element_Value_ppm),
         skewness = moments::skewness(Element_Value_ppm),
         kurtosis = moments::kurtosis(Element_Value_ppm)) %>%
  filter(Element_Symbol %in% c("REE", "HREE", "LREE", "Ag", "Cd", "Co", "Cu", "Li", "Mo", "Ni", "Pb", "Sc", "Tl", "Zn")) %>%
  select(Element_Symbol, min, max, mean, median, sd, kurtosis)

# Saving the data into .RDS format for final report purpose
saveRDS(desc_stats_flt, './Final_report/Final_report_files/desc_stats_flt.rds')
```

<br />

## Distribution Analysis

Next, we are going to analyse the distribution of each critical elements based on their PAAS normalised values. The **Figure \@ref(fig:element-dist)** provides insights regarding the spread, central tendency, and potential outliers for each critical element.

```{r element-dist, fig.cap = "Distribution of Critical Elements (Box-Plot)", out.width = "100%", fig.width = 10, fig.height = 6, fig.align = "center"}
bp1 <- alldata %>%
  filter(Element_Symbol %in% c("Ag","Al","Au","Ba","Be","Bi","Cd","Ce","Co","Cr","Cs","Cu","Dy",
                               "Er","Eu","Fe","Ga", "Gd","Ge","Ho","In","La","Li","Lu","Mn","Mo")) %>%
  ggplot(aes(x = PAAS_normalised_value, y = Element_Symbol)) +
  geom_vline(xintercept = 1, linetype = 2, color = "red") +
  xlab("PASS Normalised Value") +
  ylab("Element Symbol") +
  geom_boxplot()

bp2 <- alldata %>%
  filter(Element_Symbol %in% c("Nb","Nd","Ni","Pb","Pr","Rb","Re","Sc","Sm","Sn","Sr","Ta","Tb",
                               "Th","Tl","Tm","U","V","Y","Yb","Zn","Zr","REE","REEY","LREE","MREE","HREE")) %>%
  ggplot(aes(x = PAAS_normalised_value, y = Element_Symbol)) +
  geom_vline(xintercept = 1, linetype = 2, color = "red") +
  xlab("PASS Normalised Value") +
  ylab("Element Symbol") +
  geom_boxplot() +
  xlim(0, 45)

grid.arrange(bp1, bp2, ncol = 2)
```

As depicted above, there are some key points that we would like to raise, which are:

- **Medians**: Most of the elements have their medians close to 0, indicating that the majority of the values are low or concentrated around a lower range. Considering that the threshold to define whether a sample is below/above background is 1, some elements that have median above the threshold are: 
```{r med-above-one, fig.align = "left"}
med_above_one <- alldata %>% 
  group_by(Element_Symbol) %>% 
  summarise(median = median(PAAS_normalised_value)) %>% 
  filter(median > 1) %>%
  arrange(-median)

knitr::kable(med_above_one, 
             col.names = gsub("[_]", " ", names(med_above_one)),
             digits = 3, 
             format.args = list(big.mark = ",", scientific = FALSE),
             format = "html", 
             align = "c") %>%
  kableExtra::kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"),
                            full_width = F,
                            font_size = 14,
                            fixed_thead = T) %>%
  scroll_box(height = "200px",
             extra_css = "border: none; box-shadow: none; padding: 5px; ")
```

- **Spread and Variability**: The elements exhibit varying degrees of spread. For example, **Ge** shows a wide range with its box stretching from a low value near 0 to a higher value around 10, indicating a large variability. On the opposite side, elements like **Mn**, **Fe**, **Cr**, **Co**, **Rb**, **Ni**, **HREE** have very narrow IQR, indicating less variability.
- **Outliers**: Several elements have significant outliers, as indicated by the dots outside the whiskers of the box plots. For instance, **Ge**, **Bi**, **Mn**, and **Ba** show notable outliers far from the main data range. Additionally, some elements that was reported from **table \@ref(tab:desc-stats)**, which have high kurtosis are **Al**, **Ba**, **Bi**, **Ce**, **Co**, **Cr**, **Cs**, **Cu**, **Fe**, **Mn**, **Nb**, and **Ni**. These outliers suggest the presence of some unusually high or low values for these elements, which could be of interest for further investigation.
- **Symmetry and Skewness**: Some elements like **Ge**, **Bi**, and **Ag** appear to have a right skew, with longer whiskers or outliers extending to the right, indicating that the distribution of their normalised values has a tail on the higher end. Elements like **Ga** and **Sm** show a more symmetrical distribution with whiskers extending fairly equally on both sides of the box.
- **Comparison Accross Elements**: **Ge** stands out with a particularly large spread and median, making it an outlier among the elements. Conversely, many rare earth elements (**REE**, **REEY**, **MREE**, **LREE**, **HREE**) have relatively low medians and a small spread, indicating that their normalised values are generally low and clustered.

The plot below represents the distribution of elements from ME-4ACD81 test that will be used for predictive modelling.
```{r element-dist2, fig.cap = "Distribution of Selected Critical Elements (Box-Plot)", out.width = "100%", fig.width = 8, fig.height = 10, fig.align = "center"}
bp_alt <- alldata %>%
  filter(Element_Symbol %in% c("REE", "HREE", "LREE", "Ag", "Cd", "Co", "Cu", 
                               "Li", "Mo", "Ni", "Pb", "Sc", "Tl", "Zn")) %>%
  mutate(Element_Symbol = factor(Element_Symbol, levels = rev(c("REE", "HREE", "LREE", 
                                                            "Ag", "Cd", "Co", "Cu", 
                                                            "Li", "Mo", "Ni", "Pb", 
                                                            "Sc", "Tl", "Zn")))) %>%
  ggplot(aes(x = Element_Value_ppm, y = Element_Symbol)) +
  # geom_vline(xintercept = 1, linetype = 2, color = "red") +
  xlab("Element Value (ppm)") +
  ylab("Element Symbol") +
  geom_boxplot() +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 18),
        axis.title.x = element_text(size = 18),
        axis.title.y = element_text(size = 18))


print(bp_alt)

# Save the plot to a .png file
ggsave("./Final_report/Final_report_files/figure-latex/bp_test_a.png", plot = bp_alt, width = 8, height = 6, dpi = 300)
```


```{r element-dist3, fig.cap = "Distribution of Critical Elements (Histogram)", out.width = "100%", fig.width = 10, fig.height = 6, fig.align = "center", include = FALSE, eval = FALSE}

# TRIAL & NOT TO BE INCLUDED

hist1 <- alldata %>%
  filter(Element_Symbol %in% c("Ag","Al","Au","Ba","Be","Bi","Cd","Ce","Co","Cr","Cs","Cu","Dy",
                               "Er","Eu","Fe","Ga", "Ge", "Gd","Ho","In","La","Li","Lu","Mn","Mo")) %>%
  ggplot(aes(x = PAAS_normalised_value)) +
  geom_vline(xintercept = 1, linetype = 2, color = "red") +
  xlab("PASS Normalised Value") +
  geom_histogram() +
  facet_wrap(~ Element_Symbol)

hist2 <- alldata %>%
  filter(Element_Symbol %in% c("Nb","Nd","Ni","Pb","Pr","Rb","Re","Sc","Sm","Sn","Sr","Ta","Tb",
                               "Th","Tl","Tm","U","V","Y","Yb","Zn","Zr","REE","REEY","LREE","MREE","HREE")) %>%
  ggplot(aes(x = PAAS_normalised_value)) +
  geom_vline(xintercept = 1, linetype = 2, color = "red") +
  xlab("PASS Normalised Value") +
  geom_histogram() +
  facet_wrap(~ Element_Symbol)

grid.arrange(hist1, hist2, ncol = 1)
```

<br />

## Distribution of elements with reference to PASS levels

In this analysis, we are trying to assess all critical elements towards the PAAS standard. To begin with, we start from a high-level distribution across the two main categories that we use in identifying which elements that fall under above/below standard categories. The normalised value will be flagged as "Enriched above background" if it is above 1, while the rest will be flagged as "Below background". The **Figure \@ref(fig:fig-overall)** provide the details about this high-level distribution.

```{r fig-overall, fig.cap = "The Profile of PASS Categories", out.width = "60%", fig.align = "center", fig.height = 4}
alldata %>%
  group_by(Above_PASS_flag) %>%
  summarise(count = n()) %>%
  ungroup() %>%  # Ungroup to perform operations on the entire data frame
  mutate(Above_PASS_flag = reorder(Above_PASS_flag, count, sum)) %>%  # Reorder based on the sum of counts
  ggplot(aes(x = Above_PASS_flag, y = count, fill = Above_PASS_flag)) +
  geom_text(aes(label = count), size = 4, vjust = -0.5, color = "black", position = position_dodge(0.9)) +
  geom_col(position = "dodge") +
  xlab(NULL) +
  ylab("Count") +
  scale_fill_manual(values = c("black", "#028A0F")) +
  theme(legend.position = "none") +
  labs(fill = "Above Pass Flag")
```
As depicted in the bar chart, the distribution appears fairly balanced between the two categories, with **5,724 instances classified as "Enriched Above Background"** and **5,355 instances classified as "Below Background"**. Such a balance highlights the importance of further detailed analysis to understand the factors contributing to this distribution, the significance of enrichment in the context of the dataset, and how these elements behave under different conditions.

Moving on to the element's level, we will assess how each critical element's PAAS level, whether they are above/below background value. The Figure \@ref(fig:figelementcons) shows the profile of each sample towards this standard and their respective flags. 

```{r figelementcons, fig.cap = "Distribution of elements with reference to PASS levels' Concentration", out.width = "100%", fig.width = 10, fig.height = 8, fig.align = "center"}
alldata %>%
  ggplot(aes(x = PAAS_normalised_value, y = Element_Symbol, color = Above_PASS_flag, shape = Above_PASS_flag)) +
  geom_point() +
  geom_vline(xintercept = 1, linetype = 2) +
  xlab("PAAS Normalised Value") +
  ylab("Element Symbol") +
  scale_color_manual(values = c("black", "#028A0F")) +
  scale_shape(solid = FALSE) +
  theme(legend.position = "bottom") +
  labs(color = "Above Pass Flag")
```
As can be seen, the majority of normalised value fall within 0 to 10. Some highlight points from this plot are:

- **Ag** does not have any normalised PAAS values below background, indicating that coal waste is rich in **Ag**. 
- **HREE** concentrations are always below background value.
- The significant differences in PAAS normalised value for some elements like **Ge** and **Re** needs further investigation as to why for a particular element that was sourced from coal and coal byproducts, such big gap could happen. In the later analysis, we will reveal if they are sampled from the same project area or they are actually sampled from different project area. 
- All "below background" samples are actually just fall close to the threshold (between 0 and 1), indicating that coal and coal byproducts are highly potential as sources of these critical elements.

Furthermore, we also analyse how the distribution of the PAAS levels across various Project Area. The **Figure \@ref(fig:figprojectcons)** shows how many elements that was recorded in each project area fall into two categories.
```{r figprojectcons, fig.cap = "The Profile of Project Area with reference to PASS levels' Concentration", out.width = "80%", fig.width = 10, fig.height = 8, fig.align = "center", message = FALSE, error = FALSE}
alldata %>%
  group_by(Project_Name, Above_PASS_flag) %>%
  summarise(count = n()) %>%
  ungroup() %>%  # Ungroup to perform operations on the entire data frame
  mutate(Project_Name = reorder(Project_Name, count, sum)) %>%  # Reorder based on the sum of counts
  ggplot(aes(x = count, y = Project_Name, fill = Above_PASS_flag)) +
  geom_text(aes(label = count), size = 3, vjust = 0.5, hjust = -0.1, color = "black", position = position_dodge(0.9)) +
  geom_col(position = "dodge") +
  #facet_wrap(~ Project_Name) +
  theme(axis.text.x = element_blank()) +
  xlab("Count") +
  ylab("Project Name") +
  scale_fill_manual(values = c("black", "#028A0F")) +
  theme(legend.position = "bottom") +
  labs(fill = "Above Pass Flag")
```

Some key observations from the plot are:

- The **"Confidential_C"** project stands out with a substantial number of elements (3,358) categorised as **"Enriched Above Background"**, significantly outnumbering the **"Below Background"** category count (1,192). This suggests a considerable concentration of elements that exceeded PAAS standard in this project.
- Project such as **"Fort Cooper"**, **"Confidential_B"** and **"Confidential_A"** display a more balanced distribution between two categories, indicating a near-equal mix of elements that either meet or fall short of the enrichment criteria.
- In contrast, several projects, such as **"Wandoan"**, **"Copabella"**, **"Lake Vermont"**, **"Collinsville"**, **"Moorvalle"**, **"Newlands"**, **"Metropolitan"**, **"Rolleston"** exhibit a higher count of elements classified as **"Below Background"**. This implies that these projects have a significant proportion of elements that do not reach the enrichment threshold.
- A few projects, including **"Oaky Creek"**, and **"Unnamed"**, contribute minimally to the overall dataset, with very few elements categorised in either categories.

Lastly, let's see how is the distribution of each critical element on every project area. The **Figure \@ref(fig:element-pro-dist)** below provide this information. As the continuation of the previous analysis, in this part we will focus more on the concentration of each element in each project area. For instance, such insights that we are going to look for are elements that have wide range of concentrations in one project, elements that have above/below background in the same project area.

```{r element-pro-dist, fig.cap = "Distribution of Critical Elements by Each Project Area", out.width = "100%", fig.width = 10, fig.height = 25, fig.align = "center"}
alldata %>%
  ggplot(aes(x = PAAS_normalised_value, y = Element_Symbol, color = Above_PASS_flag, shape = Above_PASS_flag)) +
  geom_point() +
  geom_vline(xintercept = 1, linetype = 2) +
  xlab("PAAS Normalised Value") +
  ylab("Element Symbol") +
  scale_color_manual(values = c("black", "#028A0F")) +
  scale_shape(solid = FALSE) +
  facet_wrap(~ Project_Name) +
  theme(legend.position = "bottom") +
  labs(color = "Above Pass Flag")
```
As shown by above graphs, some significant insights are:

- **Collinsville**: **Co** has wide range of concentration, which make some samples fall below background, while there is one sample that has high concentration, indicating that in one project area the same element can be significantly different. 
- **Confidential_A**: the outlier of **Tl** that was seen in **figure \@ref(fig:element-dist)** and figure \@ref(fig:figelementcons) is found here, suggesting high concentration of this element in this project area. 
- **Confidential_B**: Outliers of multiple elements (**Mo**, **Li**, **Cr**, **Ba**) that were reported in **figure \@ref(fig:element-dist)** are spotted in this project area. This fact is relevant with the **figure \@ref(fig:figprojectcons)** where this project area was the source of many samples in our data. Additionally, we can spot a wide range of **Ba** concentration, in which one of the sample is categorised as "below background", while there are two samples that categorised as "above background" with very high normalised value (above 10).
- **Confidential_C**: The same conditions as previous project area can be found in this project as well, where many outliers are found here considering many samples were taken in this project. Again, **Ba** also has a wide range of concentration, a substantial difference is spotted here, where one sample has high concentration value (above 10), while the rest are clustered together between 0 & 5. 
- **Fort Cooper**: Majority of the elements' concentration are between 0 & 10. However, we can see some elements which value are beyond 10, such as **Mn**, **Ge**, **Bi**, and **Ag**. 
- **Lake Vermont**: **Ba** records significant concentration differences, where most of its sample are below background but one of them is very high (almost 18). 
- **Wandoan**: It seems **Ge** are highly concentrated in this project area, because all of its samples taken from here have concentration beyond 10. This is also revealing where all of those outliers of **Ge** are located from, as reported in **figure \@ref(fig:element-dist)**. 

<br />

## Correlation Matrix plot

```{r corplot, fig.cap = "Correlation Matrix Plot of Critical Elements", out.width = "100%", fig.width = 12, fig.height = 12, fig.align = "center"}
res_cor_cop <- alldata %>% 
  select(Project_Name,Sample_ID,Element_Symbol,PAAS_normalised_value) %>% 
  pivot_wider(names_from = Element_Symbol,values_from = c(PAAS_normalised_value))

corr_cop <- cor(res_cor_cop[,3:55], use="pairwise.complete.obs")

pg_cop <- ggcorrplot(corr_cop, 
                     method = "square", 
                     type = "lower",
                     outline.col = "white", 
                     lab = TRUE, 
                     lab_size = 2)

print(pg_cop)
```

The figure \@ref(fig:corplot) shows the relationships between various elements. Some key observations are:

- Strongly positive correlations: There are numbers of elements that have very strong relationship (above 0.95), such as **Er**-**Dy**, **Gd**-**Eu**, **Ho**-**Dy**, **Ho**-**Er**, **Pr**-**Nd**, **Sm**-**Nd**, **Tb**-**Dy**, **Tb**-**Gd**, **Tb**-**Ho**, **Yb**-**Er**, and **Yb**-**Ho**. There are also many more elements that have correlation above 0.8.
- Moderate correlations: Many elements fall between 0.6-0.8, which indicate fairly strong relationships.
- Weaker Correlations: Elements such as **Ba** show weaker correlations with most other elements, indicating less consistent co-occurrence or independent behavior within the dataset.This also the case for **Sr**.

The correlation matrix plot below represents the matrix of elements from ME-4ACD81 test to REE, HREE, and LREE. This is to understand the correlation between all elements to the target variable as part of the predictive modelling.
```{r corplot2, fig.cap = "Correlation Matrix Plot of Selected Critical Elements", out.width = "100%", fig.width = 8, fig.height = 8, fig.align = "center"}

corr_cop_cheaper <- cor(res_cor_cop[,c("Ag", "Cd", "Co", "Cu", "Li", "Mo", "Ni", "Pb", "Sc", "Tl", "Zn", "REE", "HREE", "LREE")],
                        use="pairwise.complete.obs")

# corr_cop_cheaper <- data.frame(corr_cop[c("REE", "HREE", "LREE"), c("Ag", "Cd", "Co", "Cu", "Li", "Mo", "Ni", "Pb", "Sc", "Tl", "Zn")])

pg_cop_cheaper <- ggcorrplot(corr_cop_cheaper, 
                     method = "square", 
                     type = "lower",
                     outline.col = "white", 
                     lab = TRUE, 
                     lab_size = 3)

# Save the plot to a .png file
ggsave("./Final_report/Final_report_files/figure-latex/pg_cop_cheaper.png", plot = pg_cop_cheaper, width = 8, height = 8, dpi = 300)

print(pg_cop_cheaper)
```

<br />

## Scatter Plot

In this last section, we are going to dig deeper into critical elements that have correlation value above 0.95 (as mentioned in the previous points). For context, **Er**, **Dy**, **Gd**, **Eu**, **Ho**, **Pr**, **Nd**, **Sm**, **Tb**, and **Yb** are known as **lanthanides series**. Lanthanides are a group of the first 15 f-block elements with atomic numbers from 57 to 71. In addition to yttrium, which share many similar chemical properties with the lanthanides, these elements comprise the rare earth elements (REEs) (Mattocks et al., 2021). Their high correlation values are most likely affected due to this fact.

Furthermore, Seredin (2012) suggest that the first suggestions for recovering lanthanides and yttrium (REY) as by-products from coal deposits can be traced back to 20 years ago, following the discovery of coal beds in a Russian Far East (RFE) basin that had high REY content (0.2% - 0.3%). Additional coal seams with comparable or even higher REY concentrations (up to 1.0% in ash) were identified in six coal-bearing basins across the same region. Since then, REY-rich coal has also been discovered in coal basins in various other countries. Thus, the lanthanides are found in coal and coal by-products could be possible because of their association with the materials that make up coal. 

**Correlated element (Above 0.95)**
```{r corelement, fig.cap = "Strongly Correlated Critical Elements", out.width = "100%", fig.width = 10, fig.height = 15, fig.align = "center"}
# Setting up colour palette for scatter plot legend
project_name <- c(
"Collingwood Park" = "#240807",
"Confidential_B" = "#823638", 
"Confidential_C" = "#CE6347",
"Fort Cooper" = "#BB9A7B",
"Lake Vermont" = "#A50F68",
"Meandu" = "#A89458",          
"Power plant ash" = "#DEC353",
"Unnamed" = "#D8D4AE",
"Collinsville" = "#CEC88A",
"Coppabella" = "#A5AE80",
"Metropolitan" = "#7F926D",
"Moorvale" = "#B1D0BE",
"Newlands" = "#465A50", 
"Oaky Creek" = "#528883",
"Rolleston" = "#62A29E",
"Wandoan" = "#7C8395",
"Confidential_A" = "#6C4851"  
)

# Creating scatter plots for each strongly correlated element
# Er - Dy
p1 <- res_cor_cop %>%
  ggplot(aes(x = Er, y = Dy)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Er & Dy Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Gd - Eu
p2 <- res_cor_cop %>%
  ggplot(aes(x = Gd, y = Eu)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Gd & Eu Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Ho - Dy
p3 <- res_cor_cop %>%
  ggplot(aes(x = Ho, y = Dy)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Ho & Dy Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Ho - Er
p4 <- res_cor_cop %>%
  ggplot(aes(x = Ho, y = Er)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Ho & Er Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Pr - Nd
p5 <- res_cor_cop %>%
  ggplot(aes(x = Pr, y = Nd)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Pr & Nd Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )


# Tb - Dy
p6 <- res_cor_cop %>%
  ggplot(aes(x = Tb, y = Dy)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Tb & Dy Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Tb - Gd
p7 <- res_cor_cop %>%
  ggplot(aes(x = Tb, y = Gd)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Tb & Gd Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Tb - Ho
p8 <- res_cor_cop %>%
  ggplot(aes(x = Tb, y = Ho)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Tb & Ho Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Yb - Dy
p9 <- res_cor_cop %>%
  ggplot(aes(x = Yb, y = Dy)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Yb & Dy Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Yb - Er
p10 <- res_cor_cop %>%
  ggplot(aes(x = Yb, y = Er)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Yb & Er Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Yb - Ho
p11 <- res_cor_cop %>%
  ggplot(aes(x = Yb, y = Ho)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Yb & Ho Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Sm - Nd
p12 <- res_cor_cop %>%
  ggplot(aes(x = Sm, y = Nd)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Sm & Nd Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Sm - Pr
p13 <- res_cor_cop %>%
  ggplot(aes(x = Sm, y = Pr)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  labs(title = "Sm & Pr Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=10, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Create the legend plot
legend_plot <- res_cor_cop %>%
  ggplot(aes(x = Er, y = Dy)) +
  geom_point(aes(color = Project_Name)) +
  geom_smooth(method = "lm") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  ) +
  theme(legend.position = "bottom", legend.title = element_blank()) +
  guides(color = guide_legend(override.aes = list(size = 5)))

# Extract legend using g_legend function
g_legend <- function(a.gplot){
  tmp <- ggplotGrob(a.gplot)
  leg <- tmp$grobs[[which(sapply(tmp$grobs, function(x) x$name) == "guide-box")]]
  return(leg)
}

# Store the legend
legend <- g_legend(legend_plot)

combined_plots <- arrangeGrob(arrangeGrob(p1, p2, p3, p4, p5, p6, p7, p8, p9, p10, p11, p12, p13, ncol = 3))

# grid.arrange(p1, p2, p3, p4, p5, p6, p7, p8, p9, p10, p11, p12, p13, legend, ncol = 4)

grid.arrange(combined_plots, legend, ncol = 1, heights = c(10, 1))
```

The **Figure \@ref(fig:corelement)** above depicted show correlations between these elements, additional colour dimension was added to differentiate the project areas. In summary, there are many correlations that closely align to each other, as shown by the **tight spread of the dot points around the straight line**, which are exhibited by **Pr & Nd, Gd & Eu, Ho & Er, Tb & Dy, Yb & Er, Sm & Nd, and Sm & Pr plots**. In some of the plots, there are a few data points that **deviate from the majority of the points**. These could be potential outliers or cases where the relationship slightly diverges, which are exhibited by **Er & Dy, Ho & Dy, Yb & Dy, Yb & Ho, Tb & Ho, Tb & Gd plots**. For **Sm & Nd, and Sm & Pr plots** even though they are belong to the previous group, but when we look closely from the project area point of view, we can see that the grey dots (which represents 'Wandoan') are deviating away from the linear line. This indicates the correlation between them are not strong. Let's dig deeper into these elements by utilising plots below.  

```{r corelement-wndn, fig.cap = "Correlated Critical Elements (Sm, Pr, Nd) in Wandoan", out.width = "100%", fig.width = 6, fig.height = 2, fig.align = "center"}

res_cor_cop_wndn <- alldata %>% 
  select(Project_Name,Sample_ID,Element_Symbol,PAAS_normalised_value) %>% 
  filter(Project_Name == "Wandoan",
         Element_Symbol %in% c("Sm", "Pr", "Nd")) %>%
  pivot_wider(names_from = Element_Symbol,values_from = c(PAAS_normalised_value))

corr_cop_wndn <- cor(res_cor_cop_wndn[,3:5]) # excluding REE, MREE, LREE, REEY in the corrplot

pg_cop_wndn <- ggcorrplot(corr_cop_wndn,type = "lower",
     outline.col = "white", lab = TRUE, lab_size = 2) +
     labs(title = "Correlation Matrix for Wandoan Sample") +
     theme(plot.title = element_text(size = 7, face = "bold", vjust = 10),
           legend.text = element_text(size = 6),         
           legend.key.size = unit(0.3, 'cm'))

# Sm - Nd (Wandoan Only)
p14 <- res_cor_cop %>%
  #filter(Project_Name == "Wandoan") %>%
  ggplot(aes(x = Sm, y = Nd)) +
  geom_point(aes(color = Project_Name, alpha = ifelse(Project_Name == "Wandoan", 1, 0.05))) +
  geom_smooth(method = "lm") +
  labs(title = "Sm & Nd Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=7, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

# Sm - Pr (Wandoan Only)
p15 <- res_cor_cop %>%
  #filter(Project_Name == "Wandoan") %>%
  ggplot(aes(x = Sm, y = Pr)) +
  geom_point(aes(color = Project_Name, alpha = ifelse(Project_Name == "Wandoan", 1, 0.05))) +
  geom_smooth(method = "lm") +
  labs(title = "Sm & Pr Correlation ") +
  theme(plot.title = element_text(color="black", hjust = 0.5, size=7, face="bold"),
        legend.position = "none") +
  scale_color_manual(
    breaks = names(project_name),
    values = project_name
  )

grid.arrange(pg_cop_wndn, p14, p15, ncol = 3)
```

As can be seen in the correlation matrix of **figure \@ref(fig:corelement-wndn)**, in Wandoan, the correlation between **Sm** and **Nd** is 0.95, just right at the minimum threshold that we chose. However, **Sm** and **Pr** correlation is below 0.95, which contributes to the fact that the points deviate away from the linear line. Considering their correlation value is below our threshold, it's most likely that **Sm** and **Pr** correlation from this project is excluded for the predictive modelling part.


# Reference

- Mattocks, J. A., Tirsch, J. L., & Cotruvo, J. A., Jr (2021). Determination of affinities of lanthanide-binding proteins using chelator-buffered titrations. Methods in enzymology, 651, 23–61. https://doi.org/10.1016/bs.mie.2021.01.044
- Seredin, V. V., & Dai, S. F. (2012). Coal Deposits as Potential Alternative Sources for Lanthanides and Yttrium. International Journal of Coal Geology, 94, 67-93. https://doi.org/10.1016/j.coal.2011.11.001
- Wickham H, François R, Henry L, Müller K, Vaughan D (2023). _dplyr: A Grammar of Data Manipulation_. R package
  version 1.1.4, https://github.com/tidyverse/dplyr, <https://dplyr.tidyverse.org>.
- Wickham H, Averick M, Bryan J, Chang W, McGowan LD, François R, Grolemund G, Hayes A, Henry L, Hester J, Kuhn
  M, Pedersen TL, Miller E, Bache SM, Müller K, Ooms J, Robinson D, Seidel DP, Spinu V, Takahashi K, Vaughan D,
  Wilke C, Woo K, Yutani H (2019). “Welcome to the tidyverse.” _Journal of Open Source Software_, *4*(43), 1686.
  doi:10.21105/joss.01686 <https://doi.org/10.21105/joss.01686>.
- H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
- Xie Y (2024). _knitr: A General-Purpose Package for Dynamic Report Generation in R_. R package version 1.48,
  <https://yihui.org/knitr/>.
- Yihui Xie (2015) Dynamic Documents with R and knitr. 2nd edition. Chapman and Hall/CRC. ISBN 978-1498716963.
- Yihui Xie (2014) knitr: A Comprehensive Tool for Reproducible Research in R. In Victoria Stodden, Friedrich
  Leisch and Roger D. Peng, editors, Implementing Reproducible Computational Research. Chapman and Hall/CRC. ISBN
  978-1466561595.
- Zhu H (2024). _kableExtra: Construct Complex Table with 'kable' and Pipe Syntax_. R package version 1.4.0,
  https://github.com/haozhu233/kableExtra, <http://haozhu233.github.io/kableExtra/>.
- Taiyun Wei and Viliam Simko (2024). R package 'corrplot': Visualization of a Correlation Matrix (Version 0.94).
  Available from https://github.com/taiyun/corrplot
- Kassambara A (2023). _ggcorrplot: Visualization of a Correlation Matrix using 'ggplot2'_. R package version
  0.1.4.1, <http://www.sthda.com/english/wiki/ggcorrplot-visualization-of-a-correlation-matrix-using-ggplot2>.
- Auguie B (2017). _gridExtra: Miscellaneous Functions for "Grid" Graphics_. R package version 2.3.
- Xie Y (2024). _bookdown: Authoring Books and Technical Documents with R Markdown_. R package version 0.40,
  <https://github.com/rstudio/bookdown>.
- Xie Y (2016). _bookdown: Authoring Books and Technical Documents with R Markdown_. Chapman and Hall/CRC, Boca
  Raton, Florida. ISBN 978-1138700109, <https://bookdown.org/yihui/bookdown>.
- Komsta L, Novomestky F (2022). _moments: Moments, Cumulants, Skewness, Kurtosis and Related Tests_. R package
  version 0.14.1, http://www.komsta.net/, <https://www.r-project.org>.


:::
